+++ starsky/starskytest/Services/SearchWideDateTimeTest.cs
@@ -11,6 +11,7 @@ namespace starskytest.Services
 	[TestClass]
 	public class SearchWideDateTimeTest
 	{
+
 		[TestMethod]
 		public void SearchWideDateTimeTest_LastEdited_Equal()
 		{